moved classes out directories
