+++ src/main/java/faang/school/achievement/event/ProjectEvent.java
@@ -1,4 +1,4 @@
-package faang.school.achievement.event.project;
+package faang.school.achievement.event;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
