+++ src/main/java/faang/school/achievement/handler/impl/BusinessmanAchievementHandler.java
@@ -1,6 +1,6 @@
-package faang.school.achievement.handler.impl.project;
+package faang.school.achievement.handler.impl;
 
-import faang.school.achievement.event.project.ProjectEvent;
+import faang.school.achievement.event.ProjectEvent;
 import faang.school.achievement.handler.AbstractEventHandler;
 import faang.school.achievement.service.achievement.AchievementCache;
 import faang.school.achievement.service.achievement.AchievementService;
