+++ src/main/java/faang/school/achievement/handler/impl/LibrarianAchievementHandler.java
@@ -1,6 +1,6 @@
-package faang.school.achievement.handler.impl.album;
+package faang.school.achievement.handler.impl;
 
-import faang.school.achievement.event.album.AlbumCreatedEvent;
+import faang.school.achievement.event.AlbumCreatedEvent;
 import faang.school.achievement.handler.AbstractEventHandler;
 import faang.school.achievement.service.achievement.AchievementCache;
 import faang.school.achievement.service.achievement.AchievementService;
