+++ src/main/java/faang/school/achievement/listener/impl/AlbumEventListener.java
@@ -1,7 +1,7 @@
-package faang.school.achievement.listener.impl.album;
+package faang.school.achievement.listener.impl;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import faang.school.achievement.event.album.AlbumCreatedEvent;
+import faang.school.achievement.event.AlbumCreatedEvent;
 import faang.school.achievement.handler.EventHandler;
 import faang.school.achievement.listener.AbstractEventListener;
 import org.springframework.beans.factory.annotation.Value;
